old_sdm/python/requirement.txt
config==0.4.2
Cython==0.29.12
docutils==0.14
future==0.17.1
gast==0.2.0
gdown==3.7.3
gevent==1.4.0
gin-config==0.1.4
glog==0.3.1
# gpustat==0.6.0.dev0
gpustat==0.6.0
gym==0.12.1
h5py==2.8.0
ipython==7.5.0
matplotlib==3.1.0
mesh-tensorflow==0.0.5
mistune==0.8.4
mock==2.0.0
more-itertools==4.3.0
mpmath==1.1.0
numpy==1.16.1
opencv-python==3.4.3.18
openmesh==1.1.2
Pillow==5.3.0
progressbar==2.5
protobuf==3.7.1
pycuda==2018.1.1
PyMCubes==0.0.9
# pymesh==1.0.2 # Install pymesh2 using anadonda.
pyparsing==2.2.2
pytest==3.8.2
python-dateutil==2.7.3
python-gflags==3.1.2
pytools==2018.5.2
requests==2.22.0
Rtree==0.8.3
scikit-cuda==0.5.1
scikit-image==0.15.0
scikit-learn==0.20.0
scipy==1.3.0
six==1.11.0
sklearn==0.0
spatial==0.2.0
# tensorflow-gpu==1.12.0 
trimesh==2.35.49
